src/requirements.txt
autopep8==1.6.0
blinker==1.4
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
elasticsearch==6.3.1
Flask==2.0.2
Flask-Login==0.5.0
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
Whoosh==2.7.4
